Split services.py: extract lineup.py and odds_details.py
services.py was 1,836 lines and did five unrelated jobs. Two came out with
clean seams, taking it to 1,144:

- lineup.py (194): build_lineup and build_lineup_diffs. These turned out to
  reference no module-level name in services at all -- pure functions over
  the player/defense dicts -- so the extraction is mechanical.
- odds_details.py (531): _norm_name, get_player_odds_details and
  get_defense_odds_details, backing /player/odds and /defense/odds. This one
  still needs _resolve_identity, _fetch_odds, _implied_total and
  NO_GAMES_SCHEDULED_MESSAGE from services, so it imports them. The
  dependency runs one way only and services.py must not import it back; the
  module docstring says so.

api.py now imports build_lineup/build_lineup_diffs from .lineup and calls
odds_details.* for the two detail endpoints.

Test targets moved with the code. @patch("refactored.services.<module>.<fn>")
keeps working after a move because it mutates the shared module object, but
a directly-imported function does not: the two detail-endpoint tests now
patch refactored.odds_details.odds_client. test_range_model.py imports
build_lineup from refactored.lineup rather than relying on services
re-exporting it.

Verified: ruff check + format clean, 58/58 tests pass, and GET /health and
GET / both still return 200 from the WSGI app.
