+++ tests/test_services_week_windows.py
@@ -4,14 +4,15 @@ pre-season gap (today's calendar week has no real games yet), which the
 user first mistook for a bad Odds API key. The root cause was
 weekly_windows.compute_week_windows() being anchored to today's calendar
 date with no fallback -- see test_weekly_windows.py for the fix itself.
-These tests check that every services.py call site actually uses the fix
-and surfaces a clear message instead of a bare empty list.
+These tests check that every call site actually uses the fix and surfaces a
+clear message instead of a bare empty list. The detail endpoints have since
+moved to odds_details.py, so they are exercised there.
 """
 
 import unittest
 from unittest.mock import patch
 
-from refactored import services
+from refactored import odds_details, services
 
 FAKE_ROSTER = {"players": {}, "scoring_rules": {}}
 
@@ -51,13 +52,13 @@ class NoGamesScheduledYetTest(unittest.TestCase):
         self.assertIn("message", result)
         self.assertIn("No scheduled games", result["message"])
 
-    @patch("refactored.services.odds_client.get_nfl_events")
+    @patch("refactored.odds_details.odds_client.get_nfl_events")
     def test_get_player_odds_details_surfaces_message_instead_of_bare_empty_result(
         self, mock_events
     ):
         mock_events.return_value = []
 
-        result = services.get_player_odds_details(
+        result = odds_details.get_player_odds_details(
             username="wesnicol", season="2026", week="this", name="Josh Allen"
         )
 
@@ -65,13 +66,13 @@ class NoGamesScheduledYetTest(unittest.TestCase):
         self.assertIn("message", result)
         self.assertIn("No scheduled games", result["message"])
 
-    @patch("refactored.services.odds_client.get_nfl_events")
+    @patch("refactored.odds_details.odds_client.get_nfl_events")
     def test_get_defense_odds_details_surfaces_message_instead_of_bare_empty_result(
         self, mock_events
     ):
         mock_events.return_value = []
 
-        result = services.get_defense_odds_details(
+        result = odds_details.get_defense_odds_details(
             username="wesnicol", season="2026", week="this", defense="Buffalo Bills"
         )
